refactor(notifcation): adjust component name
