+++ FinEdgeFrontend/src/App.tsx
@@ -6,7 +6,7 @@ import Login from "./Pages/Authentication/Login";
 import PersistLogin from "./Pages/Authentication/axios/Components/PersistLogin";
 import Transactions from "./Pages/Transactions/Transactions";
 import FinancialRecommendations from "./Pages/Recommendations/FinancialRecommendations";
-import NotificationsBase from "./Components/Notifications/NotificationsBas";
+import NotificationsBase from "./Components/Notifications/NotificationsBase";
 import Home from "./Pages/Home";
 import Accounts from "./Pages/Accounts";
 import Categories from "./Pages/Categories";
